services/evidence-builder/requirements.txt
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
aio-pika==9.3.1
asyncpg==0.29.0
pydantic==2.5.0